Add files via upload
The analysis in the WeatherPy section for the charts is in the commented sections after the code for each chart. 

I couldn't get the VacationPy to run. I kept getting issues saying my API key was invalid but the key worked for the WeatherPy section so I'm a little confused. I'm fine with the bad grade as long as someone can explain what I was doing wrong. Ian tried to help by showing me his code and explaining what he did but it still wouldn't work when I tried the same thing.